Set padding-right to .nick instead of span
Adjust the CSS to only apply padding-right for the .nick class, instead of every span. Ensuring no clickable spot is after the timestamps to highlight logs with.

The chances are, if we ever are to implement highlighting by nick-basis, that this will need readjusting. 

But for now:
This closes Korobi/korobi#49