Remove unnecessary requirement for trailing commas
PHP manual describes this practice as "unusual", I see it as a waste of a byte and although it improves diff output in some cases, it's not supported in some languages (e.g. JS) and it seems like a bad habit to get into. Hell, the .php_cs file itself didn't obey this rule.